Refactored runner to include timings and cancellation for long running tasks. Improved answer to Challenge 1 by handling large numbers

+++ ProjectEuler/Challenge.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; }
         public int Inputs { get; set; }
-        public int ExpectedOutput { get; set; }
+        public long ExpectedOutput { get; set; }
     }
 }
