+++ ProjectEuler/Result.cs
@@ -5,7 +5,7 @@ namespace ProjectEuler
     public class Result
     {
         public ResultState ResultState { get; set; }
-        public int ActualOutput { get; set; }
+        public long ActualOutput { get; set; }
         public TimeSpan ElapsedTime { get; set; }
     }
 }
